students/k3343/Chernomor_Margarita/Lr3/requirements.txt
Django~=5.1.3
python-dateutil~=2.9.0.post0
djangorestframework~=3.15.2
drf-yasg~=1.21.8